+++ ethergo/client/capture_client.go
@@ -3,12 +3,14 @@ package client
 import (
 	"context"
 	"fmt"
+	"net/http"
+	"time"
+
 	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/ethereum/go-ethereum/rpc"
 	"github.com/lmittmann/w3"
 	"github.com/synapsecns/sanguine/core/metrics"
 	"github.com/synapsecns/sanguine/core/metrics/instrumentation/httpcapture"
-	"net/http"
 )
 
 // captureClient is a wrapper around ethclient that can (but doesn't have to) captures requests and responses.
@@ -18,8 +20,11 @@ type captureClient struct {
 	rpcClient *rpc.Client
 }
 
+var defaultTimeout = 30 * time.Second
+
 func newCaptureClient(ctx context.Context, url string, handler metrics.Handler, capture bool) (*captureClient, error) {
 	client := new(http.Client)
+	client.Timeout = defaultTimeout
 
 	if capture {
 		client.Transport = httpcapture.NewCaptureTransport(client.Transport, handler)